+++ lib/src/utilities/global_error_handler.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import '../exceptions/exceptions.dart';
 import '../helpers/helpers.dart';
 import '../widgets/widgets.dart';
 
@@ -8,7 +9,7 @@ abstract interface class IAppErrorHandler {
   void onError();
 }
 
-/// A [GlobalErrorHandler] to ahndle uncaught errors globally. The [convertException] function should be used to check the exception and return a
+/// A [GlobalErrorHandler] to ahndle uncaught errors globally. The [handleException] function should be used to check the exception and return a
 /// nullable [ErrorData] object.
 ///
 /// This can be used with the riperpod package
@@ -29,10 +30,59 @@ abstract interface class IAppErrorHandler {
 /// ```
 ///
 /// Remember to assign the [navigationKey] to your router to allow for any [BuildContext] related error displaying.
+///
+/// This class makes use of a new exception [ControllerException] which can be extended. The [Type] source is required.
+/// Example exception class:
+/// ```dart
+/// class AppException extends ControllerException {
+///  final Type innerSource;
+///
+///  AppException(this.innerSource);
+///  @override
+///  String get message => 'You are not authorized to access the resources you requested.';
+///
+///  @override
+///  Type get source => innerSource;
+///
+///  @override
+///  String get title => 'App Error';
+/// }
+/// ```
 class GlobalErrorHandler {
+  /// Represents the collection of controller handlers that will extend [IAppErrorHandler] in order to
+  /// handle the errors globally. Useful when using providers in your application for state management
+  ///
+  /// Example using riverpod:
+  /// ```dart
+  /// class ExampleNotifier extends AsyncNotifier<ExampleState> implements IErrorHandler {...}
+  /// ```
   final Map<Type, IAppErrorHandler Function()> controllerHandlers;
+
+  /// Represends the widget (normally an alert dialog) that displays the error when an exeption happens.
+  /// One can use an overlay widget to show the error as well.
   final Future<T?> Function<T>(BuildContext context, OverlayState? overlay, ErrorData errorDetails) showErrorMessage;
-  final ErrorData? Function(Object exception) convertException;
+
+  /// This function parameter accepts an [Object] exception and converts the data to an [ErrorData] objext,
+  ///
+  /// Simplified Example:
+  /// ```dart
+  /// convertException: (error) {
+  ///     if (error is Exception) {
+  ///        return ErrorData(message: error.toString(), title: 'Error', exception: error);
+  ///     }
+  ///     if (error is ControllerException)
+  ///     {
+  ///       return ErrorData(message: error.message, title: error.title, exception: error);
+  ///     }
+  ///     return null;
+  ///  }
+  /// ```
+  final ErrorData? Function(Object exception) handleException;
+
+  /// This is useful when initialising async riverpod providers. Should return an error widget when an error occurs during
+  /// riverpod provider initialization
+  /// related to the available [handleInitRiverpodProviderError] function.
+  final Widget? Function(ErrorData data) riverpodErrorWidget;
 
   /// Use this navigator key to enable the showing of any build context related error widgets
   late LabeledGlobalKey<NavigatorState> navigationKey;
@@ -40,19 +90,25 @@ class GlobalErrorHandler {
   GlobalErrorHandler({
     required this.controllerHandlers,
     required this.showErrorMessage,
-    required this.convertException,
+    required this.handleException,
+    required this.riverpodErrorWidget,
     LabeledGlobalKey<NavigatorState>? navKey,
   }) {
     navigationKey = navKey ?? LabeledGlobalKey<NavigatorState>('coder_matthews_global_error_handler_navigator_key');
   }
+
+  /// Sets uo a [GlobalErrorHandler] with a simple alert dialog when errors occur, It is recommend that you use
+  /// the regular [GlobalErrorHandler] constructor for more customization.
   factory GlobalErrorHandler.withDefaultShowErrorDialog(
           {required Map<Type, IAppErrorHandler Function()> controllerHandlers,
-          required ErrorData? Function(Object exception) convertException,
+          required ErrorData? Function(Object exception) handleException,
+          required Widget? Function(ErrorData data) riverpodErrorWidget,
           LabeledGlobalKey<NavigatorState>? navKey}) =>
       GlobalErrorHandler(
           controllerHandlers: controllerHandlers,
           showErrorMessage: showDefaultErrorDialog,
-          convertException: convertException);
+          riverpodErrorWidget: riverpodErrorWidget,
+          handleException: handleException);
 
   /// This function should be called to handle flutter errors
   void handleFlutterError(FlutterErrorDetails details) => _handleError(details.exception, () {
@@ -73,7 +129,7 @@ class GlobalErrorHandler {
   }
 
   void _handleError(Object error, void Function() handleDefault) {
-    var errorData = convertException.call(error);
+    var errorData = handleException.call(error);
     if (errorData == null) {
       handleDefault.call();
       return;
@@ -89,5 +145,19 @@ class GlobalErrorHandler {
         debugPrintStack(stackTrace: StackTrace.current);
       }
     });
+  }
+
+  /// Use this function if you're using riverpod and handling errors when initializing the provider.
+  ///
+  /// ie:
+  /// ```dart
+  /// provider.when(...)
+  /// ```
+  ///
+  /// Throws the exception if the exception is not handled in [handleException]
+  Widget? handleInitRiverpodProviderError(Object error, StackTrace trace) {
+    var data = handleException(error);
+    if (data == null) throw error;
+    return riverpodErrorWidget(data);
   }
 }